refactor(chat): isolate turn service from IPC
